1. 改进了gui推荐和搜索的算法 2. 修改了双向链表排序的一个bug

+++ Doc.h
@@ -17,7 +17,7 @@ public:
 	CharStringLink list;
 	CharStringLink title_list;
 	AVLMap<CharString, int> title_cnt;
-	double total_words = 0;
+	int total_words = 0;
 
 	//文本内容之类的直接读取input，不占用缓存了
 public:
@@ -27,4 +27,6 @@ public:
 	Doc(CharString & str);
 	~Doc();
 };
+
+
